services/social-ingestion/requirements.txt
APScheduler==3.10.1
python-telegram-bot==20.7
praw==7.7.1
pika==1.3.1
redis==4.5.5
requests==2.31.0
python-dotenv==1.0.1
